Refactoring to make tests running
